requirements.txt

bcrypt==5.0.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Bcrypt==1.0.1      
Flask-Login==0.6.3       
Flask-SQLAlchemy==3.1.1  
greenlet==3.2.4
itsdangerous==2.2.0      
Jinja2==3.1.6
MarkupSafe==3.0.2        
SQLAlchemy==2.0.43       
typing_extensions==4.15.0
Werkzeug==3.1.3
transformers==4.30.0
huggingface_hub==0.16.4
torch
python-dotenv==1.0.0